+++ src/HearthstoneJSON.ts
@@ -1,149 +1,128 @@
-import * as https from "https";
-import * as URL from "url";
-import {IncomingMessage} from "http";
+import * as fetch from "isomorphic-fetch";
 import {StorageBackend} from "./StorageBackend";
+import NoOpStorageBackend from "./NoOpStorageBackend";
 import LocalStorageBackend from "./LocalStorageBackend";
 import CacheProxy from "./CacheProxy";
-import {Build, Locale} from "./types";
+import {Build, BuildNumber, Locale} from "./types";
+import {CardData} from "hearthstonejson";
 
 export default class HearthstoneJSON {
 
-	public defaultLocale: string = "enUS";
-	public backend: StorageBackend;
-	public cached: boolean = null;
-	public fetched: boolean = null;
-	public fallback: boolean = null;
-	public prefix: string = "hsjson-";
-	public sourceUrl: (build: Build, locale: Locale) => string;
-	public redirected: number = 0;
+	public storage: StorageBackend;
+	public storagePrefix = "hsjson-";
 
-	constructor(sourceUrl?: (build: Build, locale: Locale) => string, backend?: StorageBackend) {
-		this.sourceUrl = sourceUrl ? sourceUrl : (build: Build, locale: Locale) => "https://api.hearthstonejson.com/v1/" + build + "/" + locale + "/cards.json";
-		this.backend = backend ? backend : new CacheProxy(new LocalStorageBackend());
+	public endpoint = "https://api.hearthstonejson.com/v1/";
+	public defaultLocale = "enUS";
+
+	public cached: boolean | null = null;
+	public fetched: boolean | null = null;
+	public fallback: boolean | null = null;
+
+	constructor(storage?: StorageBackend) {
+		if (storage === null) {
+			this.storage = new NoOpStorageBackend();
+		}
+		else if (typeof storage === "undefined") {
+			this.storage = new CacheProxy(new LocalStorageBackend());
+		}
+		else if (storage) {
+			this.storage = storage;
+		}
 	}
 
-	public get(build: Build, locale: Locale, cb: (data: any[], build?: Build, locale?: string) => void): void {
-		if (typeof locale === "function" && typeof cb === "undefined") {
-			cb = locale as (data: any[]) => void;
+	public createUrl = (build: Build, locale: Locale): string => {
+		return this.endpoint + build + "/" + locale + "/cards.json";
+	};
+
+	public extractBuild = (url: string): Build => {
+		const endpointExpression = new RegExp(this.endpoint.replace(/[\/.]/g, "\\$&"));
+		const pathExpression = /((\d+)|(latest))\/[a-zA-Z]+\/cards\.json/;
+		const pattern = new RegExp("^" + endpointExpression.source + pathExpression.source + "$");
+		const matches = pattern.exec(url);
+		if (!matches) {
+			throw new Error('No build found in url "' + url + '"');
+		}
+		return matches[1] as Build;
+	};
+
+	public get(build: Build, locale?: Locale): Promise<CardData[]> {
+		if (build === "latest") {
+			return this.getLatest(locale);
+		}
+		if (!locale) {
 			locale = this.defaultLocale;
 		}
-		if (!build || isNaN(+build)) {
-			build = "latest";
-		}
-		this.cached = false;
-		if (build !== "latest") {
-			this.fetched = false;
-			this.fallback = false;
-			let key = this.generateKey(build, locale);
-			if (this.backend.has(key)) {
-				this.cached = true;
-				cb(this.backend.get(key));
-				return;
-			}
-		}
-		this.redirected = 0;
-		this.fetch(build, locale, (data: any[], receivedBuild?: Build, receivedLocale?: string) => {
-			if (!receivedBuild) {
-				receivedBuild = build;
-			}
-			if (!receivedLocale) {
-				receivedLocale = locale;
-			}
-			this.fetched = true;
-			if (!this.fallback) {
-				this.fallback = false;
-			}
-			cb(data, receivedBuild, receivedLocale);
-			if (receivedBuild !== "latest") {
-				this.backend.set(this.generateKey(receivedBuild, receivedLocale), data);
-			}
-		}, () => {
-			if (build === "latest") {
-				if (locale === this.defaultLocale) {
-					// completely failed
-					return;
-				}
-				else {
-					locale = this.defaultLocale;
-				}
-			}
-			// fallback to latest
+		return this.getSpecificBuild(build, locale).catch(() => {
+			// fall back to latest build
 			this.fallback = true;
-			this.get("latest", locale, cb);
+			return this.getLatest(locale);
 		});
 	}
 
-	public getLatest(locale: Locale, cb: (data: any[]) => void): void {
-		this.get("latest", locale, cb);
+	public getLatest(locale?: Locale): Promise<CardData[]> {
+		if (!locale) {
+			locale = this.defaultLocale;
+		}
+		return this.fetchLatestBuildNumber(locale).then((build: BuildNumber) => this.getSpecificBuild(build, locale));
+	}
+
+	protected getSpecificBuild(build: BuildNumber, locale: Locale): Promise<CardData[]> {
+		const key = this.generateKey(build, locale);
+		if (this.storage.has(key)) {
+			return this.storage.get(key);
+		}
+		return this.fetchSpecificBuild(build, locale);
+	}
+
+	protected fetchSpecificBuild(build: BuildNumber, locale: Locale): Promise<CardData[]> {
+		const headers = new Headers();
+		headers.set("content-type", "application/json");
+		return fetch(this.createUrl(build, locale), {
+			method: "GET",
+			headers: new Headers()
+		}).then((response: Response): Promise<CardData[]> => {
+			const statusCode = response.status;
+			if (statusCode !== 200) {
+				throw new Error("Expected status code 200, got " + statusCode);
+			}
+			return response.json();
+		}).then((payload: CardData[]) => {
+			const key = this.generateKey(build, locale);
+			this.storage.set(key, payload);
+			return payload;
+		});
+	}
+
+	protected fetchLatestBuildNumber(locale: Locale): Promise<BuildNumber> {
+		return fetch(this.createUrl("latest", locale), {
+			method: "HEAD",
+			redirect: "manual",
+		}).then((response: Response): BuildNumber => {
+			// we expect to be redirected
+			const statusCode = response.status;
+			if (statusCode !== 302) {
+				throw new Error("Expected status code 302, got " + statusCode);
+			}
+			// get location header
+			const headers = response.headers;
+			const location = headers.get("location");
+			if (!location) {
+				throw new Error("Expected location header in response");
+			}
+			// extract build number
+			const build = this.extractBuild(location);
+			if (isNaN(+build)) {
+				throw new Error("Expected numeric build number");
+			}
+			const buildNumber = +build;
+			return buildNumber;
+		});
 	}
 
 	protected generateKey(build: Build, locale: Locale): string {
 		if (build === "latest") {
 			throw new Error('Refusing to generate key for "latest" metadata');
 		}
-		return this.prefix + build + "_" + locale;
-	}
-
-	protected fetch(build: Build, locale: Locale, cb?: (data: any[], build: Build, locale: Locale) => void, errorCb?: () => void, url?: string): void {
-		if (typeof url === "undefined") {
-			url = this.sourceUrl(build, locale);
-		}
-		let options = URL.parse(url) as any;
-		options.withCredentials = false;
-		options.method = "GET";
-		let request = https.request(options);
-		request.once("response", (message: IncomingMessage) => {
-			if (message.statusCode != 200) {
-				if (message.statusCode >= 301 && message.statusCode <= 302 && this.redirected < 5) {
-					// redirects
-					const target = message.headers.location;
-					this.redirected++;
-					this.fetch(build, locale, cb, errorCb, target);
-					return;
-				}
-				if (message.statusCode >= 400 && message.statusCode <= 499) {
-					// bad requests
-					if (build !== "latest") {
-						// build fallback
-						this.fetch("latest", locale, cb, errorCb);
-						return;
-					}
-					if (locale !== this.defaultLocale) {
-						// locale fallback
-						this.fetch(build, this.defaultLocale, cb, errorCb);
-						return;
-					}
-					return;
-				}
-				return;
-			}
-			let data = "";
-			message.on("data", (chunk) => {
-				data += chunk;
-			});
-			message.on("error", () => {
-				errorCb();
-			});
-			message.on("end", () => {
-				try {
-					let cards = JSON.parse(data);
-					cb(cards, build, locale);
-				}
-				catch (e) {
-					errorCb();
-				}
-			});
-		});
-		request.on("error", (e) => {
-			errorCb();
-		});
-		request.end();
-	}
-
-	protected has(build: Build, locale: Locale): boolean {
-		if (build === "latest") {
-			return false;
-		}
-		return this.backend.has(this.generateKey(build, locale));
+		return this.storagePrefix + build + "_" + locale;
 	}
 }