+++ src/types.d.ts
@@ -1,2 +1,3 @@
-export type Build = number|"latest";
+export type BuildNumber = number;
+export type Build = BuildNumber|"latest";
 export type Locale = string;