+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Upgrade Typescript to 2.2.1
+- Switch to a Promise-based API using fetch
 
 ## [0.2.1] - 2016-12-14
 ### Fixed